requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.2
mysqlclient==2.1.0